apps/api/requirements.txt
annotated-types==0.7.0
anthropic==0.64.0
anyio==4.9.0
argcomplete==3.5.3
asgiref==3.8.1
async-timeout==4.0.3
attrs==18.2.0
backports.tarfile==1.2.0
build==1.2.2.post1
CacheControl==0.14.1
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
cleo==2.1.0
click==8.2.1
crashtest==0.4.1
cryptography==45.0.6
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
dulwich==0.22.7
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
fastjsonschema==2.21.1
filetype==1.2.0
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
keyring==25.6.0
langchain==0.3.27
langchain-anthropic==0.3.18
langchain-core==0.3.74
langchain-deepseek==0.1.4
langchain-google-genai==2.1.9
langchain-openai==0.3.30
langgraph==0.6.5
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.0
langsmith==0.4.14
lxml==6.0.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgpack==1.1.0
openai==1.99.9
ordered-set==4.1.0
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pdflatex==0.1.3
pkginfo==1.12.0
poetry-core==2.0.0
postgrest==1.1.1
proto-plus==1.26.1
protobuf==6.32.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.11.0
realtime==2.7.0
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.1.0
rich-toolkit==0.15.0
rignore==0.6.4
rsa==4.9.1
sentry-sdk==2.34.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.0
starlette==0.47.2
storage3==0.12.1
StrEnum==0.4.15
supabase==2.18.1
supabase_auth==2.12.3
supabase_functions==0.10.1
tenacity==9.1.2
tiktoken==0.9.0
tomli==2.2.1
tomlkit==0.13.2
tqdm==4.67.1
trove-classifiers==2024.10.21.16
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
xlsxwriter==3.2.5
xxhash==3.5.0
zstandard==0.23.0
